+++ Configuration.cs
@@ -36,6 +36,7 @@ public class TweakConfigs
     public EnhancedExpBar.Configuration EnhancedExpBar { get; init; } = new();
     public AetherCurrentHelper.Configuration AetherCurrentHelper { get; init; } = new();
     public LockWindowPosition.Configuration LockWindowPosition { get; init; } = new();
+    public MaterialAllocation.Configuration MaterialAllocation { get; init; } = new();
 }
 
 internal partial class Configuration